Updated PageResource & OptionResource (avoiding code duplication)
